--- src/input/projects/life-in-balance/data.js
@@ -1,21 +1,0 @@
-import planet from './lifeinbalance-planet.png'
-
-// images
-// IMPORT IMAGES FROM GALLERY USING THE FOLLOWING FORMAT
-// import [imageName] from './gallery/[imageName].jpg'
-
-const projectPage = [
-	{
-		tag: 'header',
-		text: 'Life In Balance'
-	},
-	{
-		tag: 'paragraph',
-		text: 'TBD'
-	}
-] 
-
-export default {
-	planet,
-	projectPage
-}
