+++ src/input/projects/lightbeing/data.js
@@ -1,0 +1,21 @@
+import planet from './lightbeing-planet.png'
+
+// images
+// IMPORT IMAGES FROM GALLERY USING THE FOLLOWING FORMAT
+// import [imageName] from './gallery/[imageName].jpg'
+
+const projectPage = [
+	{
+		tag: 'header',
+		text: 'Life In Balance'
+	},
+	{
+		tag: 'paragraph',
+		text: 'TBD'
+	}
+] 
+
+export default {
+	planet,
+	projectPage
+}
